+++ src/main/java/de/fxnm/runnable/account/service/LogOutRunnable.java
@@ -1,30 +1,28 @@
-package de.fxnm.callable.account.service;
+package de.fxnm.runnable.account.service;
 
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.project.Project;
 
-import de.fxnm.callable.BaseCallable;
+import de.fxnm.runnable.BaseRunnable;
 import de.fxnm.config.settings.password_safe.PasswordManager;
 
 import static de.fxnm.config.settings.password_safe.PasswordManager.LOGIN_DATE;
 
-public class LogOutCallable extends BaseCallable<Boolean> {
+public class LogOutRunnable extends BaseRunnable {
 
-    public LogOutCallable(final Project project) {
+    public LogOutRunnable(final Project project) {
         super(project);
     }
 
     @Override
-    public Boolean call() {
+    public void run() {
         try {
-            super.startCallable("Logging out...", "Log out started");
+            super.startRunnable("Logging out...", "Log out started");
             this.logout();
-            super.finishedCallable("Logout successful");
+            super.finishedRunnable("Logout successful");
         } catch (final Throwable e) {
-            this.failedCallable("Failed to logout");
-            return false;
+            this.failedRunnable("Failed to logout");
         }
-        return true;
     }
 
     private void logout() {