+++ src/main/java/de/fxnm/util/PooledThread.java
@@ -3,7 +3,6 @@ package de.fxnm.util;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.diagnostic.Logger;
 
-import java.util.concurrent.Callable;
 import java.util.concurrent.Future;
 
 public final class PooledThread {
@@ -14,13 +13,8 @@ public final class PooledThread {
         throw new IllegalAccessException();
     }
 
-    public static <T> Future<T> execute(final Callable<T> callable) {
-        LOG.info(callable.getClass() + " Executed");
-        return ApplicationManager.getApplication().executeOnPooledThread(callable);
-    }
-
-    public static void execute(final Runnable runnable) {
+    public static Future<?> execute(final Runnable runnable) {
         LOG.info(runnable.getClass() + " Executed");
-        ApplicationManager.getApplication().executeOnPooledThread(runnable);
+        return ApplicationManager.getApplication().executeOnPooledThread(runnable);
     }
 }